# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Gateway AI
  version: 2.0.0
  description: Gateway OpenAI Compatible para DeepSeek e aliases CLI isolados.
paths:
...
  /v1/chat/completions:
    post:
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model, messages]
              properties:
                model: { type: string, minLength: 1 }
                messages: { type: array, items: {} }
                stream: { type: boolean }
                tools: { type: array, items: {} }
                tool_choice: {}
                parallel_tool_calls: { type: boolean }
              additionalProperties: true
      responses:
        '200':
          description: Chat Completion JSON ou SSE
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
            text/event-stream:
              schema: { type: string }
        '400':
          description: Requisição inválida ou conteúdo CLI não suportado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          description: Rate limit ou broker ocupado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '502':
          description: Falha/saída inválida do provider
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503':
          description: Provider CLI indisponível
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '504':
          description: Timeout do provider
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type, code]
          properties:
            message: { type: string }
            type: { const: gateway_error }
            code: { type: string }